move doSaveState into indexer
