+++ x-pack/plugin/transform/src/main/java/org/elasticsearch/xpack/transform/transforms/ClientTransformIndexer.java
@@ -56,7 +56,6 @@ import org.elasticsearch.xpack.transform.persistence.SeqNoPrimaryTermAndIndex;
 import org.elasticsearch.xpack.transform.transforms.pivot.SchemaUtil;
 import org.elasticsearch.xpack.transform.utils.ExceptionRootCauseFinder;
 
-import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -255,119 +254,6 @@ class ClientTransformIndexer extends TransformIndexer {
             getConfig().getDestination().getIndex(),
             fieldMappingsListener
         );
-    }
-
-    @Override
-    protected void doSaveState(IndexerState indexerState, TransformIndexerPosition position, Runnable next) {
-        if (context.getTaskState() == TransformTaskState.FAILED) {
-            logger.debug("[{}] attempted to save state and stats while failed.", getJobId());
-            // If we are failed, we should call next to allow failure handling to occur if necessary.
-            next.run();
-            return;
-        }
-        if (indexerState.equals(IndexerState.ABORTING)) {
-            // If we're aborting, just invoke `next` (which is likely an onFailure handler)
-            next.run();
-            return;
-        }
-
-        // getting the listeners that registered till now, in theory a new listener could sneak in between this line
-        // and the next, however this is benign: we would store `shouldStopAtCheckpoint = true` twice which is ok
-        Collection<ActionListener<Void>> saveStateListenersAtTheMomentOfCalling = saveStateListeners.getAndSet(null);
-        boolean shouldStopAtCheckpoint = context.shouldStopAtCheckpoint();
-
-        // If we should stop at the next checkpoint, are STARTED, and with `initialRun()` we are in one of two states
-        // 1. We have just called `onFinish` completing our request, but `shouldStopAtCheckpoint` was set to `true` before our check
-        // there and now
-        // 2. We are on the very first run of a NEW checkpoint and got here either through a failure, or the very first save state call.
-        //
-        // In either case, we should stop so that we guarantee a consistent state and that there are no partially completed checkpoints
-        if (shouldStopAtCheckpoint && initialRun() && indexerState.equals(IndexerState.STARTED)) {
-            indexerState = IndexerState.STOPPED;
-            auditor.info(transformConfig.getId(), "Transform is no longer in the middle of a checkpoint, initiating stop.");
-            logger.info("[{}] transform is no longer in the middle of a checkpoint, initiating stop.", transformConfig.getId());
-        }
-
-        // This means that the indexer was triggered to discover changes, found none, and exited early.
-        // If the state is `STOPPED` this means that TransformTask#stop was called while we were checking for changes.
-        // Allow the stop call path to continue
-        if (hasSourceChanged == false && indexerState.equals(IndexerState.STOPPED) == false) {
-            if (saveStateListenersAtTheMomentOfCalling != null) {
-                ActionListener.onResponse(saveStateListenersAtTheMomentOfCalling, null);
-            }
-            next.run();
-            return;
-        }
-
-        TransformTaskState taskState = context.getTaskState();
-
-        if (indexerState.equals(IndexerState.STARTED) && context.getCheckpoint() == 1 && this.isContinuous() == false) {
-            // set both to stopped so they are persisted as such
-            indexerState = IndexerState.STOPPED;
-
-            auditor.info(transformConfig.getId(), "Transform finished indexing all data, initiating stop");
-            logger.info("[{}] transform finished indexing all data, initiating stop.", transformConfig.getId());
-        }
-
-        // If we are `STOPPED` on a `doSaveState` call, that indicates we transitioned to `STOPPED` from `STOPPING`
-        // OR we called `doSaveState` manually as the indexer was not actively running.
-        // Since we save the state to an index, we should make sure that our task state is in parity with the indexer state
-        if (indexerState.equals(IndexerState.STOPPED)) {
-            // If we are going to stop after the state is saved, we should NOT persist `shouldStopAtCheckpoint: true` as this may
-            // cause problems if the task starts up again.
-            // Additionally, we don't have to worry about inconsistency with the ClusterState (if it is persisted there) as the
-            // when we stop, we mark the task as complete and that state goes away.
-            shouldStopAtCheckpoint = false;
-
-            // We don't want adjust the stored taskState because as soon as it is `STOPPED` a user could call
-            // .start again.
-            taskState = TransformTaskState.STOPPED;
-        }
-
-        final TransformState state = new TransformState(
-            taskState,
-            indexerState,
-            position,
-            context.getCheckpoint(),
-            context.getStateReason(),
-            getProgress(),
-            null,
-            shouldStopAtCheckpoint
-        );
-        logger.debug("[{}] updating persistent state of transform to [{}].", transformConfig.getId(), state.toString());
-
-        // we might need to call the save state listeners, but do not want to stop rolling
-        persistStateWithAutoStop(state, ActionListener.wrap(r -> {
-            try {
-                if (saveStateListenersAtTheMomentOfCalling != null) {
-                    ActionListener.onResponse(saveStateListenersAtTheMomentOfCalling, r);
-                }
-            } catch (Exception onResponseException) {
-                String msg = LoggerMessageFormat.format("[{}] failed notifying saveState listeners, ignoring.", getJobId());
-                logger.warn(msg, onResponseException);
-            } finally {
-                next.run();
-            }
-        }, e -> {
-            try {
-                if (saveStateListenersAtTheMomentOfCalling != null) {
-                    ActionListener.onFailure(saveStateListenersAtTheMomentOfCalling, e);
-                }
-            } catch (Exception onFailureException) {
-                String msg = LoggerMessageFormat.format("[{}] failed notifying saveState listeners, ignoring.", getJobId());
-                logger.warn(msg, onFailureException);
-            } finally {
-                next.run();
-            }
-        }));
-    }
-
-    private void persistStateWithAutoStop(TransformState state, ActionListener<Void> listener) {
-        persistState(state, ActionListener.runBefore(listener, () -> {
-            if (state.getTaskState().equals(TransformTaskState.STOPPED)) {
-                context.shutdown();
-            }
-        }));
     }
 
     /**
